refactor of tests
